backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Cache
redis==5.0.4
aioredis==2.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML / Optimization
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
ortools==9.10.4067

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5

# Utilities
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
celery==5.4.0
kombu==5.3.7

# Monitoring
prometheus-client==0.20.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
factory-boy==3.3.0
